% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getMatchingRatesByVariable}
\alias{getMatchingRatesByVariable}
\title{Produce matching statistics by variable}
\usage{
getMatchingRatesByVariable(merged_data, data, stat_variable)
}
\description{
Produce matching statistics by variable
}
