% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getMatchingRatesByVariable_DT}
\alias{getMatchingRatesByVariable_DT}
\title{Convert list of matching statistics from 'plotMatchingByGroup' to data.frame}
\usage{
getMatchingRatesByVariable_DT(variable_list, merged_data, data)
}
\description{
Convert list of matching statistics from 'plotMatchingByGroup' to data.frame
}
