% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hismatch-package.R
\docType{package}
\name{hismatch-package}
\alias{hismatch}
\alias{hismatch-package}
\title{hismatch: Linking indvidivual-level data}
\description{
Linking indvidivual-level data.
}
\author{
\strong{Maintainer}: Eirik Berger \email{eirik.berger@gmail.com}

}
\keyword{internal}
