% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Hismatch}
\alias{Hismatch}
\title{R6 Class representing two datasets to be linked.}
\description{
Has information on the two datasets, and the details of how to link them.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Hismatch-new}{\code{Hismatch$new()}}
\item \href{#method-Hismatch-print}{\code{Hismatch$print()}}
\item \href{#method-Hismatch-addLinkingVariables}{\code{Hismatch$addLinkingVariables()}}
\item \href{#method-Hismatch-setNames}{\code{Hismatch$setNames()}}
\item \href{#method-Hismatch-createLinkingData}{\code{Hismatch$createLinkingData()}}
\item \href{#method-Hismatch-createBlocks}{\code{Hismatch$createBlocks()}}
\item \href{#method-Hismatch-setupData}{\code{Hismatch$setupData()}}
\item \href{#method-Hismatch-fuzzyMatch}{\code{Hismatch$fuzzyMatch()}}
\item \href{#method-Hismatch-executeLinking}{\code{Hismatch$executeLinking()}}
\item \href{#method-Hismatch-mergeBackInData}{\code{Hismatch$mergeBackInData()}}
\item \href{#method-Hismatch-vectorReorder}{\code{Hismatch$vectorReorder()}}
\item \href{#method-Hismatch-runMatching}{\code{Hismatch$runMatching()}}
\item \href{#method-Hismatch-getStatistics}{\code{Hismatch$getStatistics()}}
\item \href{#method-Hismatch-getMatchingRatesByVariable}{\code{Hismatch$getMatchingRatesByVariable()}}
\item \href{#method-Hismatch-getMatchingRatesByVariable_DT}{\code{Hismatch$getMatchingRatesByVariable_DT()}}
\item \href{#method-Hismatch-plotMatchingByGroup}{\code{Hismatch$plotMatchingByGroup()}}
\item \href{#method-Hismatch-iterative_link}{\code{Hismatch$iterative_link()}}
\item \href{#method-Hismatch-iterative_link_by_year}{\code{Hismatch$iterative_link_by_year()}}
\item \href{#method-Hismatch-clone}{\code{Hismatch$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hismatch-new"></a>}}
\if{latex}{\out{\hypertarget{method-Hismatch-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hismatch$new(
  data1 = NA,
  data2 = NA,
  firstname = NA,
  surname = NA,
  blocks = NA,
  dist_thr = 0.75,
  rel_thr = NA,
  max_block_size = 50000,
  letters = 1,
  matching_method = c("jw"),
  unify_middlenames = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hismatch-print"></a>}}
\if{latex}{\out{\hypertarget{method-Hismatch-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hismatch$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hismatch-addLinkingVariables"></a>}}
\if{latex}{\out{\hypertarget{method-Hismatch-addLinkingVariables}{}}}
\subsection{Method \code{addLinkingVariables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hismatch$addLinkingVariables(data_input)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hismatch-setNames"></a>}}
\if{latex}{\out{\hypertarget{method-Hismatch-setNames}{}}}
\subsection{Method \code{setNames()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hismatch$setNames(data_input, suffix)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hismatch-createLinkingData"></a>}}
\if{latex}{\out{\hypertarget{method-Hismatch-createLinkingData}{}}}
\subsection{Method \code{createLinkingData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hismatch$createLinkingData(data_input)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hismatch-createBlocks"></a>}}
\if{latex}{\out{\hypertarget{method-Hismatch-createBlocks}{}}}
\subsection{Method \code{createBlocks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hismatch$createBlocks(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hismatch-setupData"></a>}}
\if{latex}{\out{\hypertarget{method-Hismatch-setupData}{}}}
\subsection{Method \code{setupData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hismatch$setupData(data_input, suffix)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hismatch-fuzzyMatch"></a>}}
\if{latex}{\out{\hypertarget{method-Hismatch-fuzzyMatch}{}}}
\subsection{Method \code{fuzzyMatch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hismatch$fuzzyMatch(data1, data2, block_number)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hismatch-executeLinking"></a>}}
\if{latex}{\out{\hypertarget{method-Hismatch-executeLinking}{}}}
\subsection{Method \code{executeLinking()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hismatch$executeLinking(data1, data2, blocks_data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hismatch-mergeBackInData"></a>}}
\if{latex}{\out{\hypertarget{method-Hismatch-mergeBackInData}{}}}
\subsection{Method \code{mergeBackInData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hismatch$mergeBackInData(merged_data, master1, master2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hismatch-vectorReorder"></a>}}
\if{latex}{\out{\hypertarget{method-Hismatch-vectorReorder}{}}}
\subsection{Method \code{vectorReorder()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hismatch$vectorReorder(dataset)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hismatch-runMatching"></a>}}
\if{latex}{\out{\hypertarget{method-Hismatch-runMatching}{}}}
\subsection{Method \code{runMatching()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hismatch$runMatching()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hismatch-getStatistics"></a>}}
\if{latex}{\out{\hypertarget{method-Hismatch-getStatistics}{}}}
\subsection{Method \code{getStatistics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hismatch$getStatistics()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hismatch-getMatchingRatesByVariable"></a>}}
\if{latex}{\out{\hypertarget{method-Hismatch-getMatchingRatesByVariable}{}}}
\subsection{Method \code{getMatchingRatesByVariable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hismatch$getMatchingRatesByVariable(stat_variable, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hismatch-getMatchingRatesByVariable_DT"></a>}}
\if{latex}{\out{\hypertarget{method-Hismatch-getMatchingRatesByVariable_DT}{}}}
\subsection{Method \code{getMatchingRatesByVariable_DT()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hismatch$getMatchingRatesByVariable_DT(variable_list, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hismatch-plotMatchingByGroup"></a>}}
\if{latex}{\out{\hypertarget{method-Hismatch-plotMatchingByGroup}{}}}
\subsection{Method \code{plotMatchingByGroup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hismatch$plotMatchingByGroup(type_restriction)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hismatch-iterative_link"></a>}}
\if{latex}{\out{\hypertarget{method-Hismatch-iterative_link}{}}}
\subsection{Method \code{iterative_link()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hismatch$iterative_link(
  output_folder,
  block_list,
  name_string,
  source_from,
  source_to
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hismatch-iterative_link_by_year"></a>}}
\if{latex}{\out{\hypertarget{method-Hismatch-iterative_link_by_year}{}}}
\subsection{Method \code{iterative_link_by_year()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hismatch$iterative_link_by_year(
  data,
  years,
  name_string,
  output_folder,
  block_list,
  plus_years = 1
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hismatch-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Hismatch-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hismatch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
