% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotMatchingByGroup}
\alias{plotMatchingByGroup}
\title{Plot matching statistics by variable}
\usage{
plotMatchingByGroup(matching_data, type_restriction)
}
\description{
Plot matching statistics by variable
}
