% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run}
\alias{run}
\title{Run all}
\usage{
run(
  data1,
  data2,
  firstname = NULL,
  surname = NULL,
  firstname1 = NULL,
  surname1 = NULL,
  firstname2 = NULL,
  surname2 = NULL,
  dist_thr,
  rel_thr,
  blocks
)
}
\description{
Run all
}
