% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify_names.R
\name{unify_names}
\alias{unify_names}
\title{Unify the format of two names based on their middle names}
\usage{
unify_names(name1, name2)
}
\arguments{
\item{name1}{A character vector of the first names.}

\item{name2}{A character vector of the second names.}
}
\value{
A list with two elements: \code{name1} and \code{name2}, which are the modified versions of the input names.
}
\description{
Unify the format of two names based on their middle names
}
