% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{executeLinking}
\alias{executeLinking}
\title{Loop over fuzzy matching using the future package}
\usage{
executeLinking(data1, data2, blocks_data)
}
\description{
Loop over fuzzy matching using the future package
}
